Add files via upload
The flow here is as follows:
The first host wants to send files with the extension xyz to the second server, and the third server will periodically pull these files from the second server. But sometimes there are xyz files being sent from the first to the second while the third server is pulling the file. In this case, the third server can receive some xyz files in half.
To avoid this, you need to change the extension. When entering 3 scripts here in crontab, their timing should be adjusted accordingly.